Played around with some Bayesian analysis
I got a prior from the NaPiRE 2 data set (manually) and used it to calculate a posterior as well as a confidence interval (the „highest probability density (hpd)“ interval) for the question on the usage of workshops for requirements elicitiation.

So far it makes quite a lot of sense for me.